renamed App to Neighborhood to facilitate proxy server
