+++ client/src/Neighborhood.test.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 import jest from 'jest';
 import { shallow, configure } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import App from './App';
+import Neighborhood from './Neighborhood';
 
 configure({
   adapter: new Adapter()
 });
 
 test('should exist', () => {
-  const wrap = shallow(<App />);
+  const wrap = shallow(<Neighborhood />);
   expect(wrap.exists()).toEqual(true);
 });
