+++ client/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.jsx';
+import Neighborhood from './Neighborhood';
 
-ReactDOM.render(<App />, document.getElementById('app'));
+ReactDOM.render(<Neighborhood />, document.getElementById('app'));
