+++ server/index.js
@@ -1,5 +1,5 @@
 const app = require('./app.js');
 
-app.listen(8080, () => {
-  console.log('Listening on port 8080');
+app.listen(3006, () => {
+  console.log('Neighborhood: Listening on port 3006');
 });
